FIX-1937 rework ContentTable with Tanstack
