+++ Src/WitsmlExplorer.Frontend/components/ContentViews/BhaRunsListView.tsx
@@ -64,7 +64,11 @@ export const BhaRunsListView = (): React.ReactElement => {
     dispatchOperation({ type: OperationType.DisplayContextMenu, payload: { component: <BhaRunContextMenu {...contextProps} />, position } });
   };
 
-  return Object.is(selectedWellbore?.bhaRuns, bhaRuns) && <ContentTable columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />;
+  return (
+    Object.is(selectedWellbore?.bhaRuns, bhaRuns) && (
+      <ContentTable viewId="bhaRunsListView" columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />
+    )
+  );
 };
 
 export default BhaRunsListView;
