+++ Src/WitsmlExplorer.Frontend/components/ContentViews/MessagesListView.tsx
@@ -67,7 +67,11 @@ export const MessagesListView = (): React.ReactElement => {
     dispatchOperation({ type: OperationType.DisplayContextMenu, payload: { component: <MessageObjectContextMenu {...contextProps} />, position } });
   };
 
-  return Object.is(selectedWellbore?.messages, messages) && <ContentTable columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />;
+  return (
+    Object.is(selectedWellbore?.messages, messages) && (
+      <ContentTable viewId="messagesListView" columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />
+    )
+  );
 };
 
 export default MessagesListView;
