+++ Src/WitsmlExplorer.Frontend/components/ContentViews/MudLogsListView.tsx
@@ -82,7 +82,7 @@ export const MudLogsListView = (): React.ReactElement => {
 
   return (
     Object.is(selectedWellbore?.mudLogs, mudLogs) && (
-      <ContentTable columns={columns} data={getTableData()} onSelect={onSelect} onContextMenu={onContextMenu} checkableRows showRefresh />
+      <ContentTable viewId="mudLogsListView" columns={columns} data={getTableData()} onSelect={onSelect} onContextMenu={onContextMenu} checkableRows showRefresh />
     )
   );
 };
