+++ Src/WitsmlExplorer.Frontend/components/ContentViews/RigsListView.tsx
@@ -77,7 +77,8 @@ export const RigsListView = (): React.ReactElement => {
   };
 
   return (
-    selectedWellbore && Object.is(selectedWellbore.rigs, rigs) && <ContentTable columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />
+    selectedWellbore &&
+    Object.is(selectedWellbore.rigs, rigs) && <ContentTable viewId="rigsListView" columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />
   );
 };
