+++ Src/WitsmlExplorer.Frontend/components/ContentViews/RisksListView.tsx
@@ -72,7 +72,11 @@ export const RisksListView = (): React.ReactElement => {
     dispatchOperation({ type: OperationType.DisplayContextMenu, payload: { component: <RiskObjectContextMenu {...contextProps} />, position } });
   };
 
-  return Object.is(selectedWellbore?.risks, risks) && <ContentTable columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />;
+  return (
+    Object.is(selectedWellbore?.risks, risks) && (
+      <ContentTable viewId="risksListView" columns={columns} data={getTableData()} onContextMenu={onContextMenu} checkableRows showRefresh />
+    )
+  );
 };
 
 export default RisksListView;
