+++ Src/WitsmlExplorer.Frontend/components/ContentViews/WellsListView.tsx
@@ -58,7 +58,7 @@ export const WellsListView = (): React.ReactElement => {
       {wells.length > 0 && filteredWells.length == 0 ? (
         <>No wells match the current filter</>
       ) : (
-        <ContentTable columns={columns} data={getTableData()} onSelect={onSelect} onContextMenu={onContextMenu} checkableRows />
+        <ContentTable viewId="wellsListView" columns={columns} data={getTableData()} onSelect={onSelect} onContextMenu={onContextMenu} checkableRows />
       )}
     </WellProgress>
   );
